+++ src/app/app.module.ts
@@ -29,7 +29,7 @@ import { LogsModule } from '../modules/logs/logs.module';
 import { DeviceProvider } from '../providers/device/device';
 import { DataStoreProvider } from '../providers/data-store/data-store';
 import { SecureStorage } from '@ionic-native/secure-storage';
-import { TestModule } from '../modules/test/test.module';
+import { TestsModule } from '../modules/tests/tests.module';
 
 @NgModule({
   declarations: [App],
@@ -42,7 +42,7 @@ import { TestModule } from '../modules/test/test.module';
     EffectsModule.forRoot([]),
     AppInfoModule,
     LogsModule,
-    TestModule,
+    TestsModule,
   ],
   bootstrap: [IonicApp],
   entryComponents: [App],
